SIMS #171 - Student Selects 'Next Page' on Financial Information Page.
Created the place holder for the "Confirm Submission" page.
Created a default component for the footer buttons to have them aligned better to the left and right, as shown in the below image. This component has just the intention to avoid code duplication along with the pages. I don't believe that this will be the final component that we are going to be using, hence I did not spend time on unit testing it.