[web][api][SIMS #227]Adding multiple locations (#483)
Refactor the add location logic on API to allow the creations of one user with multiple permissions with only one API call.
Kept the validation logic on the controller and moved the main logic to a service (Institution service for now), but I think that this logic should be in an "institution-user" service. Also, I think that we should consider moving the API methods from the institution controller to a specific controller.