Fix Form.IO cached form load data
While loading data into the Form.IO there was a race condition between the method loading the form definition and the method loading the form data to be displayed. The race condition was already in place but became evident when the form was cache and started to be always ready before the data loading to populate the form.