[web][api][SIMS #400]Student File Upload and saved to database
Allow the student to upload files to the server and have them associated with the current Student Application.
The files will be uploaded once they are added to the Form.IO File component, and the server will create the database relationship between the files and the Student Application when the application is submitted.
Added server-side validation for
- Allowed file extensions;
- Maximum file size;
- Maximum upload parts;
Created a FileService to replace the Form.IO default one. It allows the proper authentication/authorization of the file being uploaded and downloaded. Upload and download and now handled by axios and authorized in the same way as the other API call on Vue application.
Created tables to keep the student files relationships.