Fix for CRA bypass (#821)
Fix on the CRA bypass that is configured on DEV environment to allow the CRA verification to happen immediately after the workflow is started.
Since the CRA process was moved inside a subprocess it caused a small delay that make the call from SIMS-Api to release the workflow event not be waiting for the message when the message is sent. Basically, the message is sent to the workflow even before the workflow is waiting for it.
A basic retry was added to solve it quickly.