[SIMS-#73][DB]Part1 - Migration Scripts for the SIN validation table (#1066)
Moved the student table altercolumns migration files in a single folder, checked in local doing a local-clean and running the db migrations again. It worked fine.

The TODO is just a placefolder for a future story on where to implement the sin validation when the student changes any vital information.

This table is used as auditing table for sin validation request/response file is sent and received.

For examples on the status code the match columns will get, I have created the appropriate enums for your reference.