[api][SIMS #1341]Bulk Upload of Offering Details - API Only (#1402)
Created a OfferingValidationModel to centralize all offerings validations from the different ways to manage the offerings. The OfferingValidationModel was built using the same decorators framework (class-validator) used to execute the DTO validations to try to make it easier to understand and maintain. It was also necessary the use the already in place framework class-transformer

- Created the below generic class validators:
@PeriodMaxLength
@PeriodMinLength
@HasNoPeriodOverlap
@PeriodsAreBetweenLimits
- Created the below specific offering validators:
@HasValidOfferingPeriodForFundedDays
@ProgramAllowsOfferingDelivery
@ProgramAllowsOfferingIntensity
@ProgramAllowsOfferingWIL
@StudyBreaksCombinedMustNotExceedsThreshold

- Added a package to execute the CSV parse.
- Please see below the steps necessary to have the CSV uploaded.

To be done in the upcoming PR.

- Adjust validations on SaveOfferingModel to be user-friendly;
- Test all individual validations;
- Add migrations to make SABC and Location code unique;
- Fixes on the Education Program Offering form.io definition;
- Upload UI;