#1917 - Release workflows  (#1986)
Reorganized the workflows names with the intention to have them grouped in the UI as Env Setup, Release, and Repo Checks.
- Env Setup - Build and Deploy Forms Server
- Env Setup - Deploy SIMS Secrets to Openshift
- Release - Build All
- Release - Create Branch
- Release - Deploy Camunda Resources
- Release - Deploy Form.io resources
- Repo Checks - Tests
The "build" section from .github/workflows/release-build-all.yml was not changed. The apps are building as they were before.
The semantic version to add the build using the "+" is not valid to be used as the Openshift image name and the "-" was used instead, so, instead of v1.0.0+123 we will have v1.0.0-123. The idea is still to keep a clear separation between the semantic version part and the build part. Thoughts?