#2006 - Remediate high risk vulnerabilities from WAVA report (#2019)
- Added a max value to the page number that is globally used for all API
endpoints that have pagination enabled. This fix the 45 issues on Wava
identified in the category `Integer Overflow`.
- Added an E2E test to one of the endpoints mentioned on the Wava scan:
`confirmation-of-enrollment/enrollmentPeriod/current?page=99999999999999999999&pageLimit=10`.
This E2E test case file was used because it already has tests created
and it was easier to have this one added.
- The issues on the category `API Mass Assignment` are not part of this
PR. Further details on the associated ticket.