#2017 - Upgrade npm packages to remove warnings accross all projects (#2023)
Upgraded npm packages to fix the warnings below when running `npm ci`:

![image](https://github.com/bcgov/SIMS/assets/78114138/aab8c020-c3fd-431f-8452-dd1784add78b)

### Soap issues
Updated soap version to 1.0.0 which fixed:
  - npm WARN deprecated request@2.88.2: request has been deprecated
- npm WARN deprecated har-validator@5.1.5: this library is no longer
supported

### Zeebe Node issues
@types/keyv@4.2.0 is a dependency of got@11.8.5 that is a dependency of
zeebe-node@8.1.4.

![image](https://github.com/bcgov/SIMS/assets/78114138/c7e17c39-191f-449f-8c2c-4da3cf050ab1)

The zeebee-node version won't be updated for now because of some
compatibility issues. However, even updating to zeebe-node@8.2.4 as a
test did not update `got` version and it continues displaying the
warning for `@types/keyv`. Updating `zeebe-node` to 8.2.4 updates `uuid`
and makes the warning to disappear though.

Also, we can notice that `fast-xml-parser` is a dependency of
`zeebe-node`. There are 2 dependabot alerts for this package's version:
https://github.com/bcgov/SIMS/security/dependabot/117
https://github.com/bcgov/SIMS/security/dependabot/116 
Updating `zeebe-node` to version 8.2.4 will fix them as
`fast-xml-parser` version will be 4.2.4.


### Unnecessary packages
Removed the packages below as they are not necessary any more:
  - @types/bull@4.10.0
  - @types/jwt-decode@3.1.0

### Super Test
`supertest` was updated to "^6.3.3" and removed the warning for
`superagent` and `formidable` as they are its dependencies.

### Class Transformer
Updated to version 0.5.1.

### Conclusion
There are still 2 warnings for `@types/keyv@4.2.0` and `uuid@3.4.0`. The
`uuid@3.4.0` one will be fixed when `zeebe-node` is updated to version
8.2.4 and the `@types/keyv@4.2.0` one should be fixed in some
`zeebe-node` future version.

![image](https://github.com/bcgov/SIMS/assets/78114138/649ef58c-4c52-44cf-a52b-0ed3be5fc59b)


## Info
To check the dependency tree, run:
`npm ls -a > dependencies.txt`