#2020 - Remediate high risk vulnerabilities from WAVA report (API Mass Assignment) (#2045)
### Context

Enabled the Nestjs configuration `forbidNonWhitelisted` to force the API
to throw an error and return a "bad request" if some property not mapped
to the DTO is received (more info on
[forbidNonWhitelisted](https://docs.nestjs.com/techniques/validation)).

Currently, the API already has the configuration `whitelist` which will
remove any DTO-non-mapped-property from the payload received. Enabling
the `forbidNonWhitelisted` will allow the API to return a better
response and also restrict unknown query string parameters.

#### PR Goal

The PR intention is primarily to enable the API configuration and adjust
the HTTP call from the web application to ensure that non-extra
properties will be sent. The web application sends extra values when the
payload is not "manually created" which happens basically when the
form.io submissions objects are sent directly to the API. These payloads
contain the properties needed by the API and enforced by the DTO but
they also contain some extra non-expected values that were part of the
form.io form but do not need to be part of the HTTP request.

To enforce the DTO properties, the class-transformer package was used
(the same used by Nestjs) to convert a "plain object", like the form.io
submission, to a DTO class, also removing any extra property along the
process.

Not every form.io definition had its output converted to a class because
not every form.io form is consumed in the same way.
Please see below the ones affected.

Typed to a class to exclude extraneous values.

- educationprogram
- educationprogramoffering
- institutionlocation
- institutionprofile
- institutionprofilecreation
- institutionuserprofile
- programinformationrequest
- uploadstudentdocuments
- studentprofile
- studentexceptions

No changes are needed because the payload is controlled.

- confirmsstudentenrollment
- designationagreementdetails
- reportscholasticstandingchange
- uploadstudentdocumentsaest
- exportfinancialreports
- staffapprovalappeal
- studentrequestchange

To be removed in an upcoming PR.

- approvedeclineoffering
- approvedenydesignation
- approveeducationprogram
- createnote
- declineeducationprogram
- studentapplicationdetails
- trackstudentapplication