#1954 - Request an Offering Change - Submit/View UI (Part 1) (#2105)
- Added a new endpoint to submit the request.
- Added new endpoint to get the request details submitted:
`@Get(":applicationOfferingChangeRequestId")`
- Reused methods to load programs and offerings from PIR.
- Created a component to display an offering summary to be used for the
student offering change and with the potential to be used for scholastic
standings and COE approvals.
- Created a component to submit the requested change and visualize it.
- Method `getEligibleApplications` was reused to also retrieve the
"change request"/"application details" to allow the submission. In this
way, the same query constraints are used to load the summary list and
also to get more details for the same item that will be submitted. New
endpoint added as `@Get("available/application/:applicationId")`.

### Fixes and UI refactors
- Fixed query for available applications that was showing the
application available to report a change when there was one already in
progress and also completed.
- Removed extra space from `detail-header` after the label. Instead of
showing `label : value` it will now show `label: value`.
- Replace the "vertical divider" in `detail-header` with a pipe ("|") as
per Figma.
- Fixed the `body-header` `#status-chip` that was being displayed in the
second line.