#2180 - Queued Assessments - Assessment Workflow Enqueuer Scheduler - Start Assessment (Part 1) (#2285)
- Created the new queue configuration for the scheduler, executed every
30 seconds, and cleaned up every 24 hours.
- Removed from the code the points where the queue to start the workflow
was invoked. From now on only the scheduler will invoke the workflow
start queue.
    - Method `startAssessment` removed since it is no longer needed.
- Method `hasIncompleteAssessment` and `assertAllAssessmentsCompleted`
removed since there is no purpose anymore.
- Multiple workflows can now be saved and the scheduler will coordinate
their execution.
- `StartApplicationAssessment` queue can now be started without
providing the `workflowName`.
- Created a migration to update the assessment status based on
`assessment_workflow_id`, and `assessment_data`. The rollback is less
meaningful and it was created to try to bring the column back to its
original state where the column was set with its default value
`Submitted`.