#2180 - Queue Consumers Log improvements (#2296)
Created the class `ProcessSummary` (based on the existing one) at the
generic library to allow the capture of logs alongside processes like
integrations and schedulers services. The goal is to allow the capture
of a set of logs to allow them to be logged at the same time for
convenience. For instance, the below logs are better interpreted if read
altogether.
This effort does not have the intention of removing the
`QueueProcessSummary` but it improves the logs one step further.