#2180 - Additional log improvement (#2326)
### Log improvements
- The date and time are now prepended in the job logger.
- Added a visual hierarchy to identify better the child logs.
- Changed the hierarchy organization to allow child process to be
chronologically added, as seen in the image below (before the green
ones, for instance, would be displayed altogether).

Created a new method to inspect the summary message to indicate if there
is something that needs attention in the logs.

Scheduler Enqueuer Query Change

The query was not filtering by application status and I thought that it
would be better to ensure that only `Submitted` and `Completed`
applications would be considered to have a workflow queued.

Refactor

Adjusted the logs for the `assessment-workflow-enqueuer` and
`start-application-assessment` to exclusively use the new
`ProcessSummary` class instead of the `QueueProcessSummary`. The
suggestion is intended to be a POC to have the `QueueProcessSummary`
marked as deprecated.