#2287 - Load Testing of Form.io using Student Application submission (#2407)
- Checked some possible tools for load testing,
- Chose K6 to create a POC with k6 based on the below facts:
  - It has an Open Souce version.
- Tests can be written in javascript. "JavaScript is not generally well
suited for high performance. To achieve maximum performance, the tool
itself is written in Go, embedding a JavaScript runtime allowing for
easy test scripting." (source https://k6.io/docs/).
- Out-of-box output visualization can be extended
(https://k6.io/blog/ways-to-visualize-k6-results).
  - It seems to be a well-known and largely used framework.
  - Extensive documentation and articles about it.
- Created a new package called `load-test` with two examples of tests
targeting form.io. The package was created using a merge between the
below two available templates. The main difference was that those
templates were using `babel` and I replaced them with `ts-loader` which
is the same already in place for the backend.
  - https://github.com/grafana/k6-template-typescript
  - https://github.com/grafana/k6-template-es6
- Adapted the solution to transpile typescript into javascript that K6
can execute.
- Added to the ticket comments many samples with test outputs with PODs
variation and the `Horizontal Pod Autoscaler` working.
- Currently, the `Horizontal Pod Autoscaler` in OpenShift is not working
and they will need to be adjusted. This PR is targeting the adjustment
in the forms POD only.
- Adjusted the `forms-deploy.yml` with a new configuration for the
`Horizontal Pod Autoscaler`, CPU and Memory. The CPU and memory
currently configured were not reflecting the values on DEV/TEST/PROD.
- The adjusted values, including the 50% CPU for the `Horizontal Pod
Autoscaler` were based on a 10min load test using 60 virtual users where
10334 submissions were made with 99.98% success and PODs auto-scaling
from 2 to 6 to attend the requests. The average response time was 3.49s.

### K6 considerations
- It does not execute on top of Nodejs, it is written using Go and comes
with JavaScript runtime allowing for easy test scripting, which means
that Nodejs-specific libraries won't work.

### References
- https://nodeteam.medium.com/load-testing-tools-for-node-js-developers-98291ed75a4b
- https://medium.com/@sebastian.southern/performance-testing-in-typescript-with-k6-a65377f547e6
- https://k6.io/docs/get-started/installation/
- https://k6.io/docs/
- https://k6.io/docs/using-k6/test-lifecycle/
- https://k6.io/blog/ways-to-visualize-k6-results/
- https://k6.io/docs/testing-guides/api-load-testing/
- https://levelup.gitconnected.com/load-test-your-nodejs-app-using-k6-f7b2bd8fa5cf
- https://chat.developer.gov.bc.ca/channel/devops-how-to?msg=nMAruEuBv7iL67LEu

### Documentation created (work in progress)
- https://github.com/bcgov/SIMS/wiki/Load-testing