#2322 - IER12 E2E Tests - First scenarios (#2461)
Followed a similar idea used for MSFAA E2E tests where a controlled data
set was created to allow easy file record validation with minimal
variables. So, in the same way, MSFAA is using saveMSFAATestInputData,
the IER12 E2E is using saveIER12TestInputData.

Small helper methods were created in separate files with the idea of
sharing with other E2E test files for IER12.

The payload is supposed to be expressive and clearly define the IER12
data, as shown below. Constants like `JOHN_DOE_FROM_CANADA` or
`JANE_MONONYMOUS_FROM_OTHER_COUNTRY` was created to allow the reuse of
some basic data sets.

As much as possible the IR12 models use the DB models as a reference
(e.g. using Pick<>) and the overall idea is that these models represent
the specific set of properties that would impact the IER12 file content
directly or participate in its calculations.

### First E2E scenarios for IER12.
- Should generate an IER12 file with two records for a single student
with no dependents and two disbursements, one sent and one pending.
- Should generate an IER12 file with one record for a married student
(mononymous) with dependents, applications still in assessment, and one
pending disbursement.
- Should generate an IER12 file with one record for a dependant and
living with parents student with no dependents and one sent disbursement
with no BC funding.
- Should generate an IER12 file with one record for a student with one
pending disbursement due to a restriction.

### Refactors
- Changed the student number source from application dynamic data to the
application table.