#2322 - IER12 E2E Tests - Validate Multiple Files (#2478)
- Changed the `saveIER12TestInputData` to use provided locations instead
of creating a brand new one every time.
- Once the institutions/locations were no longer brand new every time
there were conflicts with the SABC program code. The logic was adapted
to allow reusing the programs for the same institution and the same SABC
code instead of creating new programs every time (which will generate
the conflict).
- Added the scenario to allow multiple files to be generated from
different institutions/locations.
- _Should generate 2 IER12 files for locations from different
institutions using distinct institution codes._
- Created `getUploadedFiles` to allow getting multiple uploaded files.
The method `getUploadedFile` was kept for now to prevent further
refactoring and also it can be used as a helper since most of the same
we upload one file only.
- Fix for a dynamic data assertion that was being treated as a constant
which will make the E2E tests fail.