#2546 - Readiness and Liveliness Probe for workers and Queue consumers (#2550)
- [x] Have the implemented health check created from #1580 and create
readiness and liveliness probe for queue consumers
- [x] Have the implemented health check created from #1580 and create
readiness and liveliness probe for workers

Tested the scenarios with the setting mentioned in the PR for Workers
and queue consumers in dev by manually setting the liveliness and
readiness probe. Tested by making redis and/or postgres down, both the
pods restarted after the recovery successfully.