#2510 - e-Cert Implementation for Individual Student Processing - E2E Tests (#2566)
### Local Load Test
- Executed tests with 15,000 eligible disbursements and executed
refactors to improve performance and fix issues.
- Some `save` operations were converted to `update`, for performance
improvement.
- During e-Cert generation, the records were updated to `Sent` at one
single operation which for a large amount of records was leading to an
idle transaction error because the SQL command was taking too long to be
executed.

### New E2Es added for part-time/full-time.

#### Full-time

_Should disburse BC funding for a close-to-maximum disbursement, reduce
BC funding when passing the maximum, and withhold BC Funding when a
restriction was applied due to the maximum configured value for the year
being reached._

This E2E was useful to validate a restriction being added to a
disbursement and affecting the upcoming disbursement for the same
student.

#### Part-time
_Should create an e-Cert with three disbursements for two different
students with two disbursements each where three records are eligible._

#### New E2E helpers

- Created a full-time/part-time parser to alow verification of
individual fields in an e-Cert.
- Created a helper to mock the Bull Job and also allow access to check
the generated logs.

### Minor refactor
as pointed out by @guru-aot, there was a type on
`DisbursementScheduleStatus.ReadyToSend` enum. The migration file names
were not updated.