#2539 apply receive pd ppd profile verification - Part 1 (#2604)
**Acceptance Criteria** 
- [x] Suspend any ATBC integration that is already configured within
SIMS (for now)
- [x] Review the current SFAS-SIMS `pd_status` mapping logic, As per the
current logic, we have mapped `false` with `declined`. But ATBC is not
sending any declined status. Else Not Requested.
- [x] Remove Student API endpoint decorators to block ATBC API call

- As part of suspending any calls to API - commented the atbc student
controller to disable it from external calls.
- Created a new column is _active in the queue-configurations which
enables the queue to be active or not while fetching them from DB. As we
are verifying the queues in the scheduler creation and queue constants
too, commenting them will ultimately remove the queues from the bull
dashboard and disabling them.
- Skipped the test cases accordingly.