#2484 - Incorporate 'Year of Study' Field into All Offering Names (#2643)
- Added `year_of_study` to the constraint and excluded the offering type
from the constraint
conditions([#2567](https://github.com/bcgov/SIMS/issues/2567));
- Changed blurb according to business confirmation: The old banner
should be adapted to display: "The Year of Study will be displayed as
part of the offering name. Here's an example of what students will see."
and the words just below the Name field should have "Year of Study"
removed:

![image](https://github.com/bcgov/SIMS/assets/78114138/b264f425-41e4-46ca-bd72-aa569967a541)
- Displaying the Year of Study in the dropdowns:

![image](https://github.com/bcgov/SIMS/assets/78114138/4785e9ac-d9c8-4677-b9c8-73d681a3d9fe)
- Added tests for bulk upload, add offering and update offering;


### Bulk upload

![image](https://github.com/bcgov/SIMS/assets/78114138/b9714ee5-1536-432b-8499-3c05b26618db)

### Add Offering

![image](https://github.com/bcgov/SIMS/assets/78114138/11a69a33-e1a5-430f-a51a-4a7fc71c2bc8)

### Update Offering

![image](https://github.com/bcgov/SIMS/assets/78114138/2d88a154-1037-4663-85f3-db0235816626)


### Observations:
- The whole file for
[education-program-offering.model.ts](https://github.com/bcgov/SIMS/pull/2643/files#diff-50f528fd7d90633261e54ea904b6bc255ef9331a6830eae24aa5e5242024d21c)
looks different because of the wrong EOL in the file.
-
[Update-offering-duplication-unique-index.sql](https://github.com/bcgov/SIMS/pull/2643/files#diff-8ce55f89538e8f5979a8d454684e2c92dc86f587f4b1b424cd4f97ac4a589f53)
had a `commit;` at the beginning of the file because of an error with
the migration: "driverError: error: unsafe use of new value "Scholastic
standing" of enum type offering_types". That means that the
`sims.offering_types` was updated but wasn't committed. We had a
[discussion](https://github.com/bcgov/SIMS/pull/2643#pullrequestreview-1817023604)
and changed it to avoid the issue.