#2665 - Program Year Maximums Related Applications (#2732)
- Created new DB migrations for the new trigger type `Related
application changed` and the new column
`related_application_assessment_id` on `sims.student_assessments`.
- Created a centralized logic to find students' applications for the
same program year that are considered meaningful for the maximum
calculations. The new method `getSequencedApplications` will provide
information for the "past/previous" and "future" applications. This
method is prepared to handle the scope of this PR and also prepare the
logic for the upcoming ones that will require find all applications in
the past to execute a SUM of awards.
- "Past/previous" applications are the ones that should be considered
while executing the SUM of all awards already paid for the student for
that program year.
- "Future" applications are affected by a change in the current
application being modified or canceled.

### Ticket Scope
- After analysis, the Team decided that changes would generate impacts
only in the next application, which means that if there are two
applications in the future being affected, only the next one will have a
new reassessment created. Once the newly created reassessment is
processed, it will impact the next one, and so on.
- The new assessment will be created (for now) only during application
cancelation. If the application never had an assessment calculated it
will not impact any future application because no awards were ever
produced which means that no other applications took this application
into account.
- The assessment queue needed to have the allowed statuses adjusted
since now applications can have reassessments created even before they
are completed.
- The validation for saving the assessments from Camunda was also
changed to allow saving the assessment data for a reassessment before
the application is even completed. There was no apparent reason to keep
the validion or change it, hence it was removed.

#### Applications sequence order

The order of the applications in the program year to determine what
comes before an application and what comes after is determined by the
first-ever assessment calculated date for that application, including
'Overwritten' ones. 
#### Applications considered meaningful for the program year maximums
and sequence

- Current application assessment calculated.
- At least  one COE Required/Completed.