#2783 - Patroni Recovery mode Bug (#2851)
As a part of the analysis the relationId in the db models created sub
query creation, that causes DB to go into recovery mode. So removing it
will enhance the queries performance and eventually make 50K records
easily.

So as a recommendation, removing it in the all the place we use till
date is required. As a part of this effort, the relationId found that do
not have references and does not have any test case failures are
removed. The next part of this story would be removing the remaining
RelationId part of this ticket #2853