#2685 - Import Provincial Restrictions from SFAS [Bug Fix] (#2893)
### As a part of this PR, the following bug was fixed:

**Bug:** Postgres has a limitation that it can execute only one
statement with parameters as a part of the sql prepared statement. As a
result, the processing of the file `bulk-insert-restrictions.sql` from
the previous PR was failing as it had more than one sql queries all of
which were a part of the same prepared statement.

**Fix:** Both the queries of the `bulk-insert-restrictions.sql` file
have been moved into their own respective sql files and therefore, each
of them is a part of their own prepared statement. These queries form a
part of the same transaction.