#2471 -  Financial information request a change (#2901)
## Request for change - Financial Information

- [x] New form.io created for financial information - request for
change.


![image](https://github.com/bcgov/SIMS/assets/54600590/5d4b2b4d-d131-46b2-b62b-af84ca80b19c)

- [x] Request a change form enhanced with new value in the dropdown to
request for change.


![image](https://github.com/bcgov/SIMS/assets/54600590/b7aed6e0-e33e-40cd-99c2-65bb36753818)

- [x] During the load of all the change request forms(appeal forms),
program year is loaded to form initial data.
- [x] Based on the program year, the tax year in the label for the total
income is derived in a dynamic way.

![image](https://github.com/bcgov/SIMS/assets/54600590/1b1493f1-505e-48e7-aeff-9fc310163060)

- [x] All the money values entered in the form are validated against
maximum money value and negative values.

![image](https://github.com/bcgov/SIMS/assets/54600590/46adbe15-8295-4155-949c-6a6d2821e5ea)

-  [x] `MONEY_VALUE_FOR_UNKNOWN_MAX_VALUE` set to 100000000 in the code.

![image](https://github.com/bcgov/SIMS/assets/54600590/32a7c861-b4a4-4010-ba1a-574b85e1f7f1)

Note: E2E tests have been updated due to this change.
- [x] Get API updated to send the program year along with existing basic
application details.

**Note:** The existing DTO(was application identifiers DTO) has been
used as base DTO in many other application DTOs.
With the program year in it, it cannot be a base DTO anymore and hence
no more DTOs are extending this DTO.

- [x] During the dry run submission of all change request forms(appeal
forms), program year is added from the server side to perform the dry
run submission.
This action is based on the condition that, if the that particular form
requires the program year to perform dry run submission. Created a class
`StudentAppealNonSubmissionData` to type/capture all the properties of
change request forms which does not belong to the submitted data
e.g(programYear)

- [x] Added E2E tests mocking the dry run submission result but ensuring
that the dry run submission is called.


![image](https://github.com/bcgov/SIMS/assets/54600590/029bbeea-295a-4691-81aa-d591e0e7eaa3)